Dijkstra’s  Algoritham
Our project is based on dijkstra’s  algoritham .  Dijkstra's algorithm is an algorithm that is used to solve the shortest distance problem. That is, we use it to find the shortest distance between two vertices on a graph. In our project dijkstra’s algorithm works as static, initially we set a source vertex and from that source vertex the flight has to move to destination with minimum distance and it also visit the each vertex. In phase we will display our project scenario in a graph representation, which shows the source vertex ,destination vertex and the intermediate nodes present. Then we statically calculate the shortest path between the source and destination using the intermediate node values. In phase two we will represents these graph in a real scenario using openGL functions. Initially in the second phase two continents with station distance and flight  were displayed then we will assign the starting vertex ,from that staring vertex flight move to next station with minimum distance, in next phase the remaining continents with station values were displayed. From the station 2 the flight has to move to next station but the distance should be minimum ,so it select the minimum distance station ,then it move from station 2 to station 3,this process keep on going until the flight reach the destination.